Rename translationscrubber to cardMaker and make it dynamic for translations and definitions
